Add getThreads() and improve MAN
