+++ src/RcppExports.cpp
@@ -52,26 +52,6 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// cpp_get_max_thread
-int cpp_get_max_thread();
-RcppExport SEXP _proxyC_cpp_get_max_thread() {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    rcpp_result_gen = Rcpp::wrap(cpp_get_max_thread());
-    return rcpp_result_gen;
-END_RCPP
-}
-// cpp_tbb_enabled
-bool cpp_tbb_enabled();
-RcppExport SEXP _proxyC_cpp_tbb_enabled() {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    rcpp_result_gen = Rcpp::wrap(cpp_tbb_enabled());
-    return rcpp_result_gen;
-END_RCPP
-}
 // cpp_pair
 S4 cpp_pair(arma::sp_mat& mt1, arma::sp_mat& mt2, const int method, unsigned int rank, const double limit, const double weight, const double smooth, bool symm, const bool diag, const bool drop0, const bool use_nan, const int thread);
 RcppExport SEXP _proxyC_cpp_pair(SEXP mt1SEXP, SEXP mt2SEXP, SEXP methodSEXP, SEXP rankSEXP, SEXP limitSEXP, SEXP weightSEXP, SEXP smoothSEXP, SEXP symmSEXP, SEXP diagSEXP, SEXP drop0SEXP, SEXP use_nanSEXP, SEXP threadSEXP) {
@@ -94,14 +74,34 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// cpp_get_max_thread
+int cpp_get_max_thread();
+RcppExport SEXP _proxyC_cpp_get_max_thread() {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    rcpp_result_gen = Rcpp::wrap(cpp_get_max_thread());
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_tbb_enabled
+bool cpp_tbb_enabled();
+RcppExport SEXP _proxyC_cpp_tbb_enabled() {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    rcpp_result_gen = Rcpp::wrap(cpp_tbb_enabled());
+    return rcpp_result_gen;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
     {"_proxyC_cpp_linear", (DL_FUNC) &_proxyC_cpp_linear, 9},
     {"_proxyC_cpp_sd", (DL_FUNC) &_proxyC_cpp_sd, 1},
     {"_proxyC_cpp_nz", (DL_FUNC) &_proxyC_cpp_nz, 1},
+    {"_proxyC_cpp_pair", (DL_FUNC) &_proxyC_cpp_pair, 12},
     {"_proxyC_cpp_get_max_thread", (DL_FUNC) &_proxyC_cpp_get_max_thread, 0},
     {"_proxyC_cpp_tbb_enabled", (DL_FUNC) &_proxyC_cpp_tbb_enabled, 0},
-    {"_proxyC_cpp_pair", (DL_FUNC) &_proxyC_cpp_pair, 12},
     {NULL, NULL, 0}
 };
