+++ src/linear.cpp
@@ -147,24 +147,6 @@ NumericVector cpp_nz(arma::sp_mat& mt) {
     return wrap(nzs);
 }
 
-// [[Rcpp::export]]
-int cpp_get_max_thread() {
-#if PROXYC_USE_TBB
-    return tbb::this_task_arena::max_concurrency();
-#else
-    return 1;
-#endif
-}
-
-// [[Rcpp::export]]
-bool cpp_tbb_enabled(){
-#if PROXYC_USE_TBB
-    return true;
-#else
-    return false;
-#endif
-}
-
 /***R
 mt <- Matrix::rsparsematrix(100, 100, 0.01)
 system.time(
